Separate dictionary types into smaller files
