+++ libraries/dictionary/src/types/dictionary/referenceTypes.ts
