+++ libraries/dictionary/src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './dbTypes';
-export * from './dictionaryTypes';
 export * from './diffTypes';
-export * from './referenceTypes';
+export * from './dictionary';
